Modified Structure to fit Microverse Needs ? ?
